feat: add SearchFilterlistExpanded component to separate responsibility
