+++ src/components/result/filter/search-filter-list/SearchFilterList.tsx
@@ -1,22 +1,15 @@
-import { Box, HStack } from "@chakra-ui/react";
+import { Box } from "@chakra-ui/react";
 import React, { useState } from "react";
 import { MIN_MAX_BEER_VOLUME_SLIDER } from "../../../../../interface/static";
 import {
   CategoryFilterListType,
   CategoryTitle,
 } from "../../../../../interface/types";
-import { checkSelectedFilter } from "../../../../../service/filter";
-import { SearchFilterTag } from "../SearchFilterTag/SearchFilterTag";
-import {
-  SearchFilterRangeRow,
-  SearchFilterRowOption,
-  SearchFilterRowOptionsWrapper,
-  SearchFilterRowWrapper,
-} from "./SearchFilterListCell";
+import { BeerSearchCategoriesForClosedFilter } from "./SearchFilterListCell";
+import { SearchFilterListExpanded } from "./SearchFilterListExpanded";
 
 interface SearchFilterListProps {
   isFilterListOpen: boolean;
-  filterList: CategoryFilterListType[];
   selectedFilters: CategoryFilterListType[];
   onClickToggle: () => void;
   onClickTag: (targetTitle: CategoryTitle, targetTag: string) => void;
@@ -24,78 +17,25 @@ interface SearchFilterListProps {
 
 export const SearchFilterList: React.FC<SearchFilterListProps> = ({
   isFilterListOpen,
-  filterList,
   selectedFilters,
   onClickToggle,
   onClickTag,
 }) => {
-  const [beerVolume, setBeerVolume] = useState<number[]>([
-    MIN_MAX_BEER_VOLUME_SLIDER[0],
-    MIN_MAX_BEER_VOLUME_SLIDER[1],
-  ]);
   return (
     <Box>
       {isFilterListOpen ? (
-        <Box>
-          {filterList.map((filterObj) => {
-            const { title, tags, isRange } = filterObj;
-            return (
-              <SearchFilterRowWrapper key={title}>
-                <SearchFilterTag
-                  title={title}
-                  selectedFilters={selectedFilters}
-                  isFilterListOpen={isFilterListOpen}
-                  flexShrink={0}
-                />
-                <SearchFilterRowOptionsWrapper>
-                  {isRange ? (
-                    <SearchFilterRangeRow
-                      beerVolume={beerVolume}
-                      onChange={setBeerVolume}
-                    />
-                  ) : (
-                    <>
-                      {tags.map((tag: string) => {
-                        const isSelected = checkSelectedFilter(
-                          selectedFilters,
-                          title,
-                          tag
-                        );
-                        return (
-                          <SearchFilterRowOption
-                            key={tag}
-                            textColor={isSelected ? "black.100" : "gray.200"}
-                            textStyle={isSelected ? "h4_bold" : "h4"}
-                            onClick={() => {
-                              onClickTag(title, tag);
-                            }}
-                          >
-                            {tag}
-                          </SearchFilterRowOption>
-                        );
-                      })}
-                    </>
-                  )}
-                </SearchFilterRowOptionsWrapper>
-              </SearchFilterRowWrapper>
-            );
-          })}
-        </Box>
+        <SearchFilterListExpanded
+          selectedFilters={selectedFilters}
+          isFilterListOpen={isFilterListOpen}
+          onClickToggle={onClickToggle}
+          onClickTag={onClickTag}
+        />
       ) : (
-        <HStack>
-          {filterList.map((filterObj) => {
-            const { title } = filterObj;
-            return (
-              <SearchFilterTag
-                key={title}
-                title={title}
-                selectedFilters={selectedFilters}
-                onClick={onClickToggle}
-                isFilterListOpen={isFilterListOpen}
-              />
-            );
-          })}
-        </HStack>
+        <BeerSearchCategoriesForClosedFilter
+          selectedFilters={selectedFilters}
+          onClickToggle={onClickToggle}
+          isFilterListOpen={isFilterListOpen}
+        />
       )}
     </Box>
   );
